Chunk only unloads when pot broken if rose inside
Previously onBreak() didn't check if the flower pot contained a wither rose or not before unloading the chunk. Now it checks that the pot contains a wither rose so breakin empty pots or pots with other flowers in the same chunk won't interfere with the system.
boolean_1 and boolean_2 in onActivate() were renamed to better represent what they are.